# internal/web/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: harbrr management API
  version: 0.1.0
  description: >
    harbrr's HTTP surface: the management API (first-run setup, authentication,
    API-key management, indexer and app-sync configuration) plus the entry points of
    the *arr-facing Torznab feed. The Torznab wire contract itself — the XML body
    shapes — is defined by Jackett/Newznab and served by internal/torznab; here we
    document only its HTTP routes, parameters, and content types (see
    docs/architecture.md). Hand-authored, embedded via //go:embed, and guarded by a
    drift test that asserts the spec matches the mounted routes (make test-openapi).

    Authentication: the management API uses a cookie session (after login) or the
    X-API-Key header; the Torznab feed authenticates with an apikey (or passkey)
    query parameter, as *arr clients expect. CSRF: a cookie-authenticated mutating
    request (POST/PUT/PATCH/DELETE) must echo the session's CSRF token in an
    X-CSRF-Token header — the token is minted at login and handed to the client via
    the non-HttpOnly harbrr_csrf cookie and GET /api/auth/me; a missing or wrong
    token is 403. X-API-Key and auth-disabled callers are exempt (see
    docs/security.md).
  license:
    name: GPL-2.0-or-later
    url: https://www.gnu.org/licenses/gpl-2.0.html
...
paths:
...
  /api/apps:
    get:
      summary: List first-class app identities (credentials redacted)
      description: >
        Apps hold an external service's identity + one sealed credential + its harbrr
        vantage, stored once and referenced by the app-sync, announce, and download
        surfaces (ADR 0004). Each entry carries per-surface reference counts.
      operationId: listApps
      tags: [Apps]
      security: [{ session: [] }, { apiKey: [] }]
      responses:
        "200":
          description: the configured apps (credentials redacted)
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/App" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/apps/{id}:
    get:
      summary: Get one app (credential redacted)
      operationId: getApp
      tags: [Apps]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: the app
          content:
            application/json:
              schema: { $ref: "#/components/schemas/App" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      summary: Update an app (a new apiKey rotates the credential for every surface)
      operationId: updateApp
      tags: [Apps]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateApp" }
      responses:
        "204": { description: updated }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      summary: Delete an app (409 when still referenced by a surface)
      operationId: deleteApp
      tags: [Apps]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "204": { description: deleted }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/apps/{id}/qui-instances:
    get:
      summary: Proxy a qui app's instance list (server-side; the key never leaves harbrr)
      operationId: appQuiInstances
      tags: [Apps]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: the qui instances (ok=false carries a scrubbed error)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuiInstances" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: harbrr_session
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        A bearer credential — send over HTTPS only; it must never traverse a
        cleartext link. Terminate TLS at harbrr or a trusted reverse proxy.
  responses:
    BadRequest:
      description: invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: authentication required or failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: conflicting state
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, description: human-readable message }
        code:
          type: string
          description: >
            machine-readable code clients can branch on — one of: bad_request,
            invalid, unauthorized, invalid_credentials, invalid_api_key, forbidden,
            not_found, conflict, already_setup, not_implemented, internal. An OIDC
            ID-token verification failure also uses invalid_credentials.
      required: [error, code]
...
    App:
      type: object
      description: >
        A first-class external service harbrr connects to — a (kind, base_url) identity
        plus one sealed credential and the app's harbrr vantage, stored once and
        referenced by the app-sync/announce/download surfaces (ADR 0004). The credential
        (apiKey) is never echoed — it reads back as the <redacted> sentinel.
      properties:
        id: { type: integer, format: int64 }
        kind: { type: string }
        name: { type: string }
        baseUrl: { type: string }
        username: { type: string, description: empty for API-key apps; set for user+password apps }
        apiKey: { type: string, description: "the credential (API key or password), redacted (<redacted>)" }
        harbrrUrl: { type: string, description: how this app reaches harbrr's feed (unused by download clients) }
        enabled: { type: boolean }
        references:
          type: object
          description: how many surface rows reference this app
          properties:
            appConnections: { type: integer }
            announce: { type: integer }
            download: { type: integer }
          required: [appConnections, announce, download]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, kind, name, baseUrl, username, apiKey, harbrrUrl, enabled, references, createdAt, updatedAt]
    UpdateApp:
      type: object
      description: >
        A partial update; omitted fields are unchanged. apiKey, when present, rotates the
        credential — every referencing surface decrypts through the app, so the rotation
        propagates. Omit apiKey to keep the stored credential (never re-submit <redacted>).
      properties:
        name: { type: string }
        baseUrl: { type: string }
        username: { type: string }
        harbrrUrl: { type: string }
        enabled: { type: boolean }
        apiKey: { type: string }
    QuiInstance:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
      required: [id, name]
    QuiInstances:
      type: object
      description: >
        The proxied qui instance list. ok=false carries a credential-scrubbed error
        (reachability/auth failure), mirroring the test endpoints; the HTTP status is 200.
      properties:
        ok: { type: boolean }
        error: { type: string }
        instances: { type: array, items: { $ref: "#/components/schemas/QuiInstance" } }
      required: [ok]
...